Declare y and residual as fp16 too, which is what the kernel reads and writes
Same class as the routing_w fix one round earlier, in the same kernel: it declares 'device const half* residual' and
'device half* y', and both were Float buffers. The kernel wrote halfs into a Float y and the readback read floats - so
every peer answer was garbage, which is node1's NaN. The residual survived by accident, because zero is zero in both
widths; y did not.

In FP16 terms: 1.0f is 0x3F800000, which reads as the halfs 0.0 and 1.875. Nothing on the Swift side asserts what
the Metal declaration asserts, so both mistakes compiled, passed every gate, and were silently about something else.

Suite 0 failures, tools/lint.sh clean.

+++ sources/TinyTitan/Runtime/Inference/RealForwardRunner+ShardServe.swift
@@ -76,8 +76,9 @@ extension RealForwardRunner {
             // `await completed()`, NOT `waitUntilCompleted()`: Swift marks the blocking form unavailable from an
             // asynchronous context, which is the compiler saying a cooperative-pool thread must not be parked.
             await cb.completed()
-            let src = remoteY.contents().bindMemory(to: Float.self, capacity: dims)
-            for d in 0..<dims { out[index * dims + d] = src[d] }
+            // `half`, matching the kernel's `device half* y`.
+            let src = remoteY.contents().bindMemory(to: Float16.self, capacity: dims)
+            for d in 0..<dims { out[index * dims + d] = Float(src[d]) }
         }
         return out
     }
